# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Finance Management API
  description: API for financial operations including deposits, loans, and user management
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        type:
          type: string
          enum: [standard, installment]
        amount:
          type: number
          format: float
        term:
          type: integer
        interest_rate:
          type: number
          format: float
        total_payable:
          type: number
          format: float
        monthly_payment:
          type: number
          format: float
        status:
          type: string
          enum: [pending, approved, active, completed, rejected]
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        type:
          type: string
        amount:
          type: number
          format: float
        metadata:
          type: string
        timestamp:
          type: string
          format: date-time
        cancelled:
          type: boolean
        cancelled_by:
          type: integer
          format: int64
        cancel_time:
          type: string
          format: date-time
...
paths:
...
  /manager/loans/pending:
    get:
      summary: Get loans that need manager review
      tags:
        - Manager
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, active, completed]
            default: pending
      responses:
        '200':
          description: List of loans with the specified status
          content:
            application/json:
              schema:
                type: object
                properties:
                  loans:
                    type: array
                    items:
                      type: object
                      allOf:
                        - $ref: '#/components/schemas/Loan'
                        - type: object
                          properties:
                            username:
                              type: string
                            needs_review:
                              type: boolean
                  total_pending:
                    type: integer
        '401':
          description: Unauthorized - invalid token
        '403':
          description: Forbidden - manager privileges required
...
  /manager/transactions:
    get:
      summary: Get transaction history with filters
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized - invalid token
        '403':
          description: Forbidden - insufficient privileges
...
  /manager/transactions/cancel:
    post:
      summary: Cancel a transaction
      tags:
        - Transactions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transaction_id
              properties:
                transaction_id:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Transaction cancelled successfully
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized - invalid token
        '403':
          description: Forbidden - insufficient privileges